+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jsondiscovery",
-  "version": "1.5.2",
+  "version": "1.6.0",
   "description": "DiscoveryJson",
   "author": "exsdis@gmail.com",
   "license": "MIT",
