+++ safari/JsonDiscovery/JsonDiscovery/AppDelegate.swift
@@ -1,0 +1,22 @@
+//
+//  AppDelegate.swift
+//  JsonDiscovery
+//
+//  Created by Колесников Денис on 04/08/2019.
+//  Copyright © 2019 exdis.me. All rights reserved.
+//
+
+import Cocoa
+
+@NSApplicationMain
+class AppDelegate: NSObject, NSApplicationDelegate {
+    
+    func applicationDidFinishLaunching(_ aNotification: Notification) {
+        // Insert code here to initialize your application
+    }
+    
+    func applicationWillTerminate(_ aNotification: Notification) {
+        // Insert code here to tear down your application
+    }
+    
+}
