+++ safari/JsonDiscovery/JsonDiscovery/ViewController.swift
@@ -1,0 +1,30 @@
+//
+//  ViewController.swift
+//  JsonDiscovery
+//
+//  Created by Колесников Денис on 04/08/2019.
+//  Copyright © 2019 exdis.me. All rights reserved.
+//
+
+import Cocoa
+import SafariServices.SFSafariApplication
+
+class ViewController: NSViewController {
+
+    @IBOutlet var appNameLabel: NSTextField!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.appNameLabel.stringValue = "JsonDiscovery";
+    }
+    
+    @IBAction func openSafariExtensionPreferences(_ sender: AnyObject?) {
+        SFSafariApplication.showPreferencesForExtension(withIdentifier: "me.exdis.JsonDiscovery-Extension") { error in
+            if let _ = error {
+                // Insert code to inform the user that something went wrong.
+
+            }
+        }
+    }
+
+}
