+++ src/content/inject-safari.js
@@ -1,0 +1,5 @@
+import { init, getSettingsSafari } from './init';
+
+document.addEventListener('DOMContentLoaded', () => {
+    init(getSettingsSafari);
+});
